незначительные доработки
