+++ js/modal.js
@@ -1,30 +1,8 @@
 import { selectors, refreshSelectors } from './selectors.js';
+import { toggleBodyScroll } from './util.js';
 
-let scrollPosition = 0; // Текущая позиция прокрутки страницы
 let currentPhoto = null;
 let displayedCommentsCount = 0; // Счётчик отображенных комментариев
-
-// Функция для блокировки и разблокировки прокрутки страницы
-function toggleBodyScroll(isLocked) {
-  if (isLocked) {
-    scrollPosition = window.scrollY;
-    const scrollBarWidth = window.innerWidth - document.documentElement.clientWidth;
-    document.body.style.position = 'fixed';
-    document.body.style.top = `-${scrollPosition}px`;
-    document.body.style.left = '0';
-    document.body.style.width = '100%';
-    document.body.style.paddingRight = `${scrollBarWidth}px`;
-    document.body.classList.add('modal-open');
-  } else {
-    document.body.style.position = '';
-    document.body.style.top = '';
-    document.body.style.left = '';
-    document.body.style.width = '';
-    document.body.style.paddingRight = '';
-    window.scrollTo(0, scrollPosition);
-    document.body.classList.remove('modal-open');
-  }
-}
 
 // Функция для открытия полноразмерного изображения
 function openBigPicture(photo) {
